Vine branch

+++ casa-site/app/Http/Controllers/NewsletterController.php
@@ -2,10 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 use App\Newsletter;
 use Validator;
 use App\Http\Requests\NewsletterRequest;
+use App\Notifications\NewsletterAdicionadaNotification;
 
 class NewsletterController extends Controller
 {
@@ -32,23 +35,30 @@ class NewsletterController extends Controller
         $dados['receber_eventos'] = true;
         $dados['receber_projetos'] = true;
         $dados['receber_noticias'] = true;
-
+        $dados['token'] = str_replace('/', '', bcrypt(Str::random(8)));
         
-        $this->newsletter->create($dados);
-        return redirect()->route('site.home')->with('success', 'Newsletter adicionada com sucesso!');
+        $newsletter = $this->newsletter->create($dados);
+
+        $newsletter->notify(new NewsletterAdicionadaNotification($newsletter));
+
+        return redirect()->back()->with('newsletter', 'Cadastro feito com sucesso, agora você ficará sabendo das novidades!');
 
     }
 
 
     // Método responsavel por abrir a pagina de editar um projeto
-    public function editar($id) 
+    public function editar($id, $token) 
     {
         $registro = $this->newsletter->find($id);
+        if($token != $registro->token) {
+            throw new ModelNotFoundException;
+        } 
+
         return view('admin.newsletter.editar', compact('registro'));
     }
 
     // Método responsavel por salvar as informacoes do formulario de edicao no banco de dados
-    public function atualizar(Request $request,$id) 
+    public function atualizar(Request $request, $id, $token) 
     {
         $dados = $request->all();
 
@@ -71,15 +81,33 @@ class NewsletterController extends Controller
             $dados['receber_noticias'] = false;
         }
         
-        $this->newsletter->find($id)->update($dados);
-        return redirect()->route('admin.newsletters')->with('success', 'Newsletter atualizada com sucesso!');
+        $newsletter = $this->newsletter->find($id);
+        
+        if($token != $newsletter->token) {
+            throw new ModelNotFoundException;
+        } 
+        
+        $newsletter->update($dados);
+        
+        if(!$dados['receber_eventos'] && !$dados['receber_projetos'] && !$dados['receber_noticias']) {
+            $this->newsletter->find($id)->delete();
+            return redirect()->back()->with('success', 'Notificações de email canceladas com sucesso!');
+        }
+
+        return redirect()->back()->with('success', 'Notificações de email atualizadas com sucesso!');
     }
 
     // Metodo da acao de apagar uma newsletter
-    public function deletar($id) 
+    public function deletar($id, $token) 
     {
-        $this->newsletter->find($id)->delete();
-        return redirect()->route('admin.newsletters')->with('success', 'Newsletter deletada com sucesso!');
+        $registro = $this->newsletter->find($id);
+        if($token != $registro->token) {
+            throw new ModelNotFoundException;
+        } 
+
+        $registro->delete();
+
+        return redirect()->back()->with('success', 'Feito! Você não receberá mais novos email.');
     }
 }
